+++ bookops_worldcat/query.py
@@ -3,20 +3,25 @@
 """
 Handles actual requests to OCLC services
 """
-from typing import Optional, Union, Tuple
+from __future__ import annotations
+from typing import Optional, Union, Tuple, TYPE_CHECKING
 import sys
 
-from requests import Session
 from requests.models import PreparedRequest
 from requests.exceptions import ConnectionError, HTTPError, Timeout
-
 
 from .errors import WorldcatRequestError
 
 
+if TYPE_CHECKING:
+    from .metadata_api import MetadataSession  # pragma: no cover
+
+
 class Query:
     """
-    Sends a request to OClC service and unifies received excepitons
+    Sends a request to OClC service and unifies received exceptions
+    Query object handles refreshing expired token before request is
+    made to the web service.
 
     `Query.response` attribute is `requests.Response` instance that
     can be parsed to exctract received information from the web service.
@@ -24,7 +29,7 @@ class Query:
 
     def __init__(
         self,
-        session: Session,
+        session: MetadataSession,
         prepared_request: PreparedRequest,
         timeout: Optional[
             Union[int, float, Tuple[int, int], Tuple[float, float]]
@@ -32,7 +37,7 @@ class Query:
     ) -> None:
         """
         Args:
-            session:                        `requests.Session` instance
+            session:                        `metadata_api.MetadataSession` instance
             prepared_request:               `requests.models.PreparedRequest` instance
             timeout:                        how long to wait for server to send data
                                             before giving up
@@ -43,6 +48,10 @@ class Query:
         """
         if not isinstance(prepared_request, PreparedRequest):
             raise AttributeError("Invalid type for argument 'prepared_request'.")
+
+        # make sure access token is still valid and if not request a new one
+        if session.authorization.is_expired():
+            session._get_new_access_token()
 
         self.response = None
 
@@ -62,7 +71,9 @@ class Query:
                 self.response.raise_for_status()
 
         except HTTPError as exc:
-            raise WorldcatRequestError(f"{exc}")
+            raise WorldcatRequestError(
+                f"{exc}. Server response: {self.response.content}"
+            )
         except (Timeout, ConnectionError):
             raise WorldcatRequestError(f"Connection Error: {sys.exc_info()[0]}")
         except: