+++ tests/test_authorize.py
@@ -34,9 +34,7 @@ class TestWorldcatAccessToken:
             ),
         ],
     )
-    def test_key_exceptions(
-        self, argm, expectation, msg, mock_successful_post_token_response
-    ):
+    def test_key_exceptions(self, argm, expectation, msg):
         with expectation as exp:
             WorldcatAccessToken(
                 key=argm,
@@ -67,9 +65,7 @@ class TestWorldcatAccessToken:
             ),
         ],
     )
-    def test_secret_exceptions(
-        self, argm, expectation, msg, mock_successful_post_token_response
-    ):
+    def test_secret_exceptions(self, argm, expectation, msg):
         with expectation as exp:
             WorldcatAccessToken(
                 key="my_key",
@@ -80,7 +76,7 @@ class TestWorldcatAccessToken:
             )
             assert msg in str(exp.value)
 
-    def test_agent_exceptions(self, mock_successful_post_token_response):
+    def test_agent_exceptions(self):
         with pytest.raises(WorldcatAuthorizationError) as exp:
             WorldcatAccessToken(
                 key="my_key",
@@ -117,9 +113,7 @@ class TestWorldcatAccessToken:
             ),
         ],
     )
-    def test_principal_id_exception(
-        self, arg, expectation, msg, mock_successful_post_token_response
-    ):
+    def test_principal_id_exception(self, arg, expectation, msg):
         with expectation as exc:
             WorldcatAccessToken(
                 key="my_key",
@@ -145,9 +139,7 @@ class TestWorldcatAccessToken:
             ),
         ],
     )
-    def test_principal_idns_exception(
-        self, arg, expectation, msg, mock_successful_post_token_response
-    ):
+    def test_principal_idns_exception(self, arg, expectation, msg):
         with expectation as exc:
             WorldcatAccessToken(
                 key="my_key",
@@ -183,9 +175,7 @@ class TestWorldcatAccessToken:
             ),
         ],
     )
-    def test_scope_exceptions(
-        self, argm, expectation, msg, mock_successful_post_token_response
-    ):
+    def test_scope_exceptions(self, argm, expectation, msg):
         with expectation as exp:
             WorldcatAccessToken(
                 key="my_key",
@@ -354,42 +344,23 @@ class TestWorldcatAccessToken:
         )
         assert token.is_expired() is False
 
-    def test_is_expired_true(
-        self, mock_utcnow, mock_credentials, mock_successful_post_token_response
-    ):
-        creds = mock_credentials
-        token = WorldcatAccessToken(
-            key=creds["key"],
-            secret=creds["secret"],
-            scopes=creds["scopes"],
-            principal_id=creds["principal_id"],
-            principal_idns=creds["principal_idns"],
-        )
-        token.token_expires_at = datetime.datetime.strftime(
+    def test_is_expired_true(self, mock_utcnow, mock_token):
+        mock_token.is_expired() is False
+        mock_token.token_expires_at = datetime.datetime.strftime(
             datetime.datetime.utcnow() - datetime.timedelta(0, 1),
             "%Y-%m-%d %H:%M:%SZ",
         )
 
-        assert token.is_expired() is True
+        assert mock_token.is_expired() is True
 
     @pytest.mark.parametrize(
         "arg,expectation",
         [(None, pytest.raises(TypeError)), ("20-01-01", pytest.raises(ValueError))],
     )
-    def test_is_expired_exception(
-        self, arg, expectation, mock_credentials, mock_successful_post_token_response
-    ):
-        creds = mock_credentials
-        token = WorldcatAccessToken(
-            key=creds["key"],
-            secret=creds["secret"],
-            scopes=creds["scopes"],
-            principal_id=creds["principal_id"],
-            principal_idns=creds["principal_idns"],
-        )
-        token.token_expires_at = arg
+    def test_is_expired_exception(self, arg, expectation, mock_token):
+        mock_token.token_expires_at = arg
         with expectation:
-            token.is_expired()
+            mock_token.is_expired()
 
     def test_post_token_request(
         self,
@@ -414,6 +385,16 @@ class TestWorldcatAccessToken:
         assert token.scopes == "scope1 scope2"
         assert token.server_response.json() == mock_oauth_server_response.json()
         assert token.timeout == (3, 3)
+
+    def test_token_repr(
+        self,
+        mock_token,
+        mock_utcnow,
+    ):
+        assert (
+            str(mock_token)
+            == "access_token: 'tk_Yebz4BpEp9dAsghA7KpWx6dYD1OZKWBlHjqW', expires_at: '2020-01-01 17:19:58Z'"
+        )
 
     @pytest.mark.webtest
     def test_cred_in_env_variables(self, live_keys):