+++ tests/test_query.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from contextlib import nullcontext as does_not_raise
+import datetime
 import os
 
 import pytest
@@ -41,6 +42,21 @@ def test_query_not_prepared_request(stub_session):
         req = Request("GET", "https://foo.org")
         Query(stub_session, req, timeout=2)
     assert "Invalid type for argument 'prepared_request'." in str(exc.value)
+
+
+@pytest.mark.http_code(200)
+def test_query_with_stale_token(stub_session, mock_utcnow, mock_session_response):
+    stub_session.authorization.token_expires_at = datetime.datetime.strftime(
+        datetime.datetime.utcnow() - datetime.timedelta(0, 1),
+        "%Y-%m-%d %H:%M:%SZ",
+    )
+    assert stub_session.authorization.is_expired() is True
+
+    req = Request("GET", "http://foo.org")
+    prepped = stub_session.prepare_request(req)
+    query = Query(stub_session, prepped)
+    assert stub_session.authorization.is_expired() is False
+    assert query.response.status_code == 200
 
 
 @pytest.mark.http_code(200)